Title: 합과 차의 나눗셈, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240209/합과 차의 나눗셈/divide-of-sum-and-sub.cpp
@@ -7,7 +7,7 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int a, b;
     cin >> a >> b;
-    float res = static_cast<float>(a + b) / (a - b);
+    double res = static_cast<double>(a + b) / (a - b);
     res = round(res * 100.0) / 100.0;
     cout << res;
     return 0;
